test-requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --no-emit-trusted-host --output-file test-requirements.txt test-requirements.in
#
certifi==2018.11.29       # via requests
chardet==3.0.4            # via requests
click==7.0                # via flask
flask==1.0.2
flexmock==0.10.3
idna==2.8                 # via requests
itsdangerous==1.1.0       # via flask
jabstract==0.1.3
jinja2==2.10              # via flask
markupsafe==1.1.1         # via jinja2
mockserver-friendly-client==0.2.1
nose==1.3.7
pyhamcrest==1.9.0
requests==2.21.0
six==1.12.0               # via pyhamcrest
urllib3==1.24.1           # via requests
werkzeug==0.14.1          # via flask
